Reimplement capture device interface
The capture device interface is now more generic, using key-value pairs
instead of functions to communicate between VCS and the device. This
allows VCS to be decoupled more strongly from Datapath's capture devices
and so (in theory) make it easier to add other kinds of capture sources.

This reimplementation liberally culls the number of superfluous capture
features. For example, the UI no longer provides metadata about the
capture device. Sone of these features might be put back in later.

The new interface hasn't been tested on an actual capture card yet and
so is expected to require some fixing in future commits.

+++ src/anti_tear/anti_tearer.cpp
@@ -56,10 +56,10 @@ u8* anti_tearer_c::process(u8 *const pixels,
 
     anti_tear_frame_s frame(resolution, pixels);
 
-    const unsigned long minValidRowIdx = 0;
-    const unsigned long maxValidRowIdx = (frame.resolution.h - 1);
+    const unsigned minValidRowIdx = 0;
+    const unsigned maxValidRowIdx = (frame.resolution.h - 1);
     this->scanEndRow = std::max(minValidRowIdx, std::min((frame.resolution.h - this->scanEndOffset - 1), maxValidRowIdx));
-    this->scanStartRow = std::min((unsigned long)(this->scanEndRow), std::min(maxValidRowIdx, (unsigned long)(this->scanStartOffset)));
+    this->scanStartRow = std::min(this->scanEndRow, std::min(maxValidRowIdx, this->scanStartOffset));
 
     if (this->scanDirection == anti_tear_scan_direction_e::up)
     {
